backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
weasyprint==60.2
requests==2.31.0
openai==1.3.7
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
pydantic-settings==2.1.0
